fix: reverts async reducer and resolves infinite effect
